+++ xplat/src/main/java/dev/emi/emi/network/CommandS2CPacket.java
@@ -5,7 +5,7 @@ import dev.emi.emi.api.recipe.EmiRecipe;
 import dev.emi.emi.api.stack.EmiStack;
 import dev.emi.emi.bom.BoM;
 import dev.emi.emi.registry.EmiCommands;
-import dev.emi.emi.registry.EmiShareRecipe;
+import dev.emi.emi.runtime.EmiShareRecipe;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.network.RegistryByteBuf;
